# example/doc/app/app.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "app/app.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BlogService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/articles": {
      "get": {
        "operationId": "BlogService_GetArticles",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appGetArticlesResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "title",
            "description": "@inject_tag: form:\"title\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "@inject_tag: form:\"page\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "description": "@inject_tag: form:\"page_size\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "authorId",
            "description": "@inject_tag: form:\"author_id\" uri:\"author_id\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    },
    "/v1/author/{authorId}/articles": {
      "get": {
        "operationId": "BlogService_GetArticles2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appGetArticlesResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "authorId",
            "description": "@inject_tag: form:\"author_id\" uri:\"author_id\"",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "title",
            "description": "@inject_tag: form:\"title\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "@inject_tag: form:\"page\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "description": "@inject_tag: form:\"page_size\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BlogService"
        ]
      },
      "post": {
        "operationId": "BlogService_CreateArticle",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appArticle"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "authorId",
            "description": "@inject_tag: form:\"author_id\" uri:\"author_id\"",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "content",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    }
  },
  "definitions": {
    "appArticle": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "authorId": {
          "type": "integer",
          "format": "int32",
          "title": "@inject_tag: form:\"author_id\" uri:\"author_id\""
        }
      }
    },
    "appGetArticlesResp": {
      "type": "object",
      "properties": {
        "total": {
          "type": "string",
          "format": "int64"
        },
        "articles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/appArticle"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}